backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
httpx
python-dotenv
authlib
pyjwt
pydantic-settings
google-generativeai
supabase
